Restructured __init__.py loading of configurations
