+++ src/__init__.py
@@ -1,2 +1,3 @@
 # Force load_env.py to be imported early, so that env is loaded right away
-from . import load_env
+from .init import load_env
+from .init import config_pd
